requirements.txt
appdirs==1.4.4
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cssselect==1.1.0
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-heroku==0.3.1
et-xmlfile==1.1.0
fake-useragent==0.1.11
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
lxml==4.8.0
mysqlclient==2.1.0
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.1
parse==1.19.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pyee==8.2.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.63.0
trio==0.20.0
trio-websocket==0.9.2
twilio==7.7.0
urllib3==1.26.8
w3lib==1.22.0
websockets==10.2
whitenoise==6.0.0
wsproto==1.1.0
zipp==3.7.0
